Move base url implementation to impl package
