+++ app/src/main/java/gq/kirmanak/mealient/data/baseurl/VersionInfo.kt
