+++ app/src/main/java/gq/kirmanak/mealient/data/baseurl/impl/VersionResponse.kt
@@ -1,4 +1,4 @@
-package gq.kirmanak.mealient.data.baseurl
+package gq.kirmanak.mealient.data.baseurl.impl
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
