+++ app/src/main/java/gq/kirmanak/mealient/data/baseurl/impl/VersionService.kt
@@ -1,4 +1,4 @@
-package gq.kirmanak.mealient.data.baseurl
+package gq.kirmanak.mealient.data.baseurl.impl
 
 import retrofit2.http.GET
