+++ app/src/main/java/gq/kirmanak/mealient/extensions/RemoteToLocalMappings.kt
@@ -1,7 +1,7 @@
 package gq.kirmanak.mealient.extensions
 
 import gq.kirmanak.mealient.data.baseurl.VersionInfo
-import gq.kirmanak.mealient.data.baseurl.VersionResponse
+import gq.kirmanak.mealient.data.baseurl.impl.VersionResponse
 import gq.kirmanak.mealient.data.recipes.db.entity.RecipeEntity
 import gq.kirmanak.mealient.data.recipes.db.entity.RecipeIngredientEntity
 import gq.kirmanak.mealient.data.recipes.db.entity.RecipeInstructionEntity
